requirements.txt
flask
gunicorn
google-generativeai
textblob
transformers
torch
tensorflow==2.12.0  # Explicit TensorFlow version